refactor(#120): move clingen/gencc/cosmic-cgc constants to per-plugin shared/constants.py
Three plugins constants relocated in one batched PR per user choice:

  - clingen (6 constants)
  - gencc (4 constants)
  - cosmic_cgc (5 constants, including COSMIC_CGC_FILE_PREFIX which is
    currently unused; moved conservatively because the name is plugin-
    scoped, matching the move-plugin-named/delete-generic-named precedent)

cosmic_cgc gains a new shared/ directory (matching the convention used
by 8 other in-tree plugins). Dropped 2 module-import logger.debug dumps
(clingen, gencc; cosmic_cgc had none). Updated 12 Python consumers and
6 SKILL.md path references. Three of the nine PRs closing #120.
