+++ hvantk/skills/clingen/drift_probe.py
@@ -18,7 +18,7 @@ from datetime import datetime, timezone
 
 import requests
 
-from hvantk.core.constants import CLINGEN_BASE_URL, CLINGEN_FILE_PREFIX
+from hvantk.skills.clingen.shared.constants import CLINGEN_BASE_URL, CLINGEN_FILE_PREFIX
 from hvantk.core.plugin.api import DriftProbeError
 
 PROBE_VERSION = 1
