+++ hvantk/skills/gencc/drift_probe.py
@@ -17,7 +17,7 @@ from datetime import datetime, timezone
 
 import requests
 
-from hvantk.core.constants import GENCC_BASE_URL, GENCC_FILE_PREFIX
+from hvantk.skills.gencc.shared.constants import GENCC_BASE_URL, GENCC_FILE_PREFIX
 from hvantk.core.plugin.api import DriftProbeError
 
 PROBE_VERSION = 1
